requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.1
django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
markdown==3.4.3
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.6
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.7.1 ; python_version < '3.11'
